test(throw): refactor and improve throw() tests
Join the two files throw-spec.ts and ErrorObservable-spec.ts into throw-spec.ts, and add a test case
for the purpose of generating a PNG diagram.
