+++ spec/observables/throw-spec.ts
@@ -1,10 +1,18 @@
 import {expect} from 'chai';
-import * as Rx from '../../dist/cjs/Rx';
-
+import * as Rx from '../../dist/cjs/Rx.KitchenSink';
+declare const {hot, asDiagram, expectObservable, expectSubscriptions};
+import {ErrorObservable} from '../../dist/cjs/observable/ErrorObservable';
+declare const rxTestScheduler: Rx.TestScheduler;
 const Observable = Rx.Observable;
 
 /** @test {throw} */
 describe('Observable.throw', () => {
+  asDiagram('throw(e)')('should create a cold observable that just emits an error', () => {
+    const expected = '#';
+    const e1 = Observable.throw('error');
+    expectObservable(e1).toBe(expected);
+  });
+
   it('should emit one value', (done: MochaDone) => {
     let calls = 0;
     Observable.throw('bad').subscribe(() => {
@@ -14,5 +22,24 @@ describe('Observable.throw', () => {
       expect(err).to.equal('bad');
       done();
     });
+  });
+
+  it('should create expose a error property', () => {
+    const e = Observable.throw('error');
+
+    expect(e['error']).to.equal('error');
+  });
+
+  it('should create ErrorObservable via static create function', () => {
+    const e = new ErrorObservable('error');
+    const r = ErrorObservable.create('error');
+
+    expect(e).to.deep.equal(r);
+  });
+
+  it('should accept scheduler', () => {
+    const e = Observable.throw('error', rxTestScheduler);
+
+    expectObservable(e).toBe('#');
   });
 });
